backend/pyproject.toml
[project]
name = "menu-ai-backend"
version = "0.1.0"
description = "AI-powered menu parsing and recommendation backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "google-genai>=0.3.0",
    "requests>=2.31.0",
    "python-multipart>=0.0.6",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "httpx>=0.25.0",
]
